+++ CleanupSandbox/function.json
@@ -1,25 +1,11 @@
 {
   "bindings": [
     {
-      "name": "timer",
-      "type": "timerTrigger",
+      "type": "queueTrigger",
       "direction": "in",
-      "schedule": "0 */5 * * * *",
-      "runOnStartup": false
-    },
-    {
-      "name": "lastDate",
-      "type": "blob",
-      "path": "ap2020bot/CleanupSandbox",
-      "connection": "AzureWebJobsStorage",
-      "direction": "in"
-    },
-    {
-      "name": "$return",
-      "type": "blob",
-      "path": "ap2020bot/CleanupSandbox",
-      "connection": "AzureWebJobsStorage",
-      "direction": "out"
+      "name": "queue",
+      "queueName": "DumpSandbox",
+      "connection": "AzureWebJobsStorage"
     }
   ],
   "scriptFile": "../dist/CleanupSandbox/index.js"
