+++ DumpSandbox/for-cleaner.ts
@@ -1,0 +1,5 @@
+export interface QueueMessage {
+    lastDumped: {
+        ts: string;
+    };
+};
